Add files via upload
This file includes analyzing user interactions and engagement with the Instagram app to provide valuable insights that can help the business grow .User analysis involves tracking how users engage with a digital product, such as a software application or a mobile app. The insights derived from this analysis can be used by various teams within the business. For example, the marketing team might use these insights to launch a new campaign, the product team might use them to decide on new features to build, and the development team might use them to improve the overall user experience.

This Data Analytics project involves the use of SQL and MySQL Workbench.
The goal of this project is to use your SQL skills to extract meaningful insights from the data.