Add files via upload
This project report involves the use of Exploratory Data Analysis (EDA) to analyze patterns in the data and ensure that capable applicants are not rejected.

When a customer applies for a loan, the company faces two risks:

If the applicant can repay the loan but is not approved, the company loses business.
If the applicant cannot repay the loan and is approved, the company faces a financial loss.

The dataset contains information about loan applications. It includes two types of scenarios:

Customers with payment difficulties: These are customers who had a late payment of more than X days on at least one of the first Y installments of the loan.
All other cases: These are cases where the payment was made on time.

When a customer applies for a loan, there are four possible outcomes:

Approved: The company has approved the loan application.
Cancelled: The customer cancelled the application during the approval process.
Refused: The company rejected the loan.
Unused Offer: The loan was approved but the customer did not use it.

The goal is to use EDA to understand how customer attributes and loan attributes influence the likelihood of default.